// supabase/functions/_utils/types.ts
export interface Database {
  public: {
    Tables: {
      admins: {
        Row: {
          user_id: string;
        };
        Insert: {
          user_id: string;
        };
        Update: {
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "admins_user_id_fkey";
            columns: ["user_id"];
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
        ];
      };
      chat_messages: {
        Row: {
          attachments: string[];
          chat: number;
          content: string;
          created_at: string;
          id: number;
          owner: string;
          reply: number | null;
        };
        Insert: {
          attachments?: string[];
          chat: number;
          content: string;
          created_at?: string;
          id?: number;
          owner?: string;
          reply?: number | null;
        };
        Update: {
          attachments?: string[];
          chat?: number;
          content?: string;
          created_at?: string;
          id?: number;
          owner?: string;
          reply?: number | null;
        };
        Relationships: [
          {
            foreignKeyName: "chat_messages_chat_fkey";
            columns: ["chat"];
            referencedRelation: "chats";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "chat_messages_owner_fkey";
            columns: ["owner"];
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "chat_messages_reply_fkey";
            columns: ["reply"];
            referencedRelation: "chat_messages";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      profiles: {
        Row: {
          avatar_url: string | null;
          chat_notification: boolean;
          deadline_notification: boolean;
          full_name: string | null;
          id: string;
          notification_token: string | null;
          order_notification: boolean;
          updated_at: string | null;
        };
        Insert: {
          avatar_url?: string | null;
          chat_notification?: boolean;
          deadline_notification?: boolean;
          full_name?: string | null;
          id: string;
          notification_token?: string | null;
          order_notification?: boolean;
          updated_at?: string | null;
        };
        Update: {
          avatar_url?: string | null;
          chat_notification?: boolean;
          deadline_notification?: boolean;
          full_name?: string | null;
          id?: string;
          notification_token?: string | null;
          order_notification?: boolean;
          updated_at?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "profiles_id_fkey";
            columns: ["id"];
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      is_admin: {
        Args: Record<PropertyKey, never>;
        Returns: boolean;
      };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
}